🔒️ Upgrade update-notifier & other minor versions.
The update-notifier@5.1.0 package depends on a version of [got](https://github.com/sindresorhus/got) which has a security vulnerability. Upgrading to update-notifier@6.0.0 fixes this problem. Since this is a breaking change, I checked the [release notes](https://github.com/yeoman/update-notifier/releases/tag/v6.0.0) for update-notifier@6.0.0. Version 6.0.0 of update-notifier is a breaking change because it requires Node.js version 14, and because the package is now a pure ESM package. This major version upgrade is all right because the gitmoji-cli package.json file lists node version `>=14`, and because the file importing update-notifier is an ESM file.

I also generated a new `yarn.lock` file with upgraded minor versions for several packages. This fixed all the remaining security vulnerabilities reported by `yarn audit`. After that, I ran the test suite locally to ensure that all the tests are still passing.