:white_check_mark: Transform node_modules and import.meta.url
Most of the tests were failing because we were trying to import ESM
node_modules in CJS. To fix this, I set the transformIgnorePatterns to
an empty array and moved the babel config to a separate babel.config.cjs
file. This ensured that babel transformed all the node_modules before
running Jest.

The chalk packages uses the package.json imports field, which Jest does
not understand. The workaround was to use moduleNameMapper to map these
imports to the correct paths.

The src/cli.js file uses import.meta which can not be used in CJS files.
To fix this issue, I installed the babel-plugin-transform-import-meta
package and included it in the babel test environment config. This
package transforms import.meta.url into a CJS equivalent. However, it
does not transform import.meta. Hence, I manually created a meta object
in the source file with the url field set to import.meta.url.

I also set the mock implementation of fs.readFileSync to the actual
implementation of fs.readFileSync so that we can read the package.json
file from src/cli.js. In addition, I updated the snapshots as required.