:sparkles: Set default emoji format to emoji
The default emoji format was originally set to `code` for the following
reason:

> The reason behind this was to be able to use the shortcode to favour
> the readability on certain terminals that do not support Emojis!
> @carloscuesta

Terminals are considerably more advanced on all platforms and the
standard Gitmoji set is simple enough that all terminals would display
them correctly nowadays.

This change updates the default format from `code` to `emoji`.

The benefits of this change:

- Improve display of emojis in commit titles, pull request titles and
  list of pull requests on some remote repository providers.

- Removes need to modify the pager to allow `git log` to display
  emojis.

Issue: #1247
Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>